+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Memento/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Memento/releases)
 
-- Added: When a screenshot is taken, a timestamp can also be displayed in the chat [all versions]
+- Added: TOC version for patch 12.0.5 [retail]
+- Added: When a screenshot is taken, the current timestamp can also be displayed in the chat